+++ pages/index.js
@@ -60,37 +60,7 @@ export default function PageHome() {
           paragraph="Ergo-Lend started for the ErgoHack 2 hackathon and we just kept going!  We hope to help overcome some issues for the billions of unbanked people around the world.  This is some of what we have in various stages of implementation."
         />
         <Timeline />
-        <section className="border-y border-[#F3F4F4] pt-10 pb-20 dark:border-[#2D2C4A] mt-20">
-          <SectionHeader
-            className="mt-32"
-            primaryText="OUR TEAM"
-            heading="Meet Our Core Team"
-            paragraph="Ergo-Lend is a community driven tool and series of smart contracts and is not owned by anyone - but our core team contribues as part of this community effort to help organize financial communities where there are no banking options are available"
-          />
-          <MeetTheTeam />
-        </section>
-        <section className="bg-light-bg pt-[80px] pb-20 dark:bg-[#14102C]">
-          <SectionHeader
-            className="mt-32"
-            primaryText="OUR TEAM"
-            heading="Our Global Partners"
-            paragraph="Ergo-Lend community partners around the world will begin to work on initiatives to help reach the people who need access to services that were previously only offered by banks"
-          />
-          <GlobalTeam />
-        </section>
       </MainLayout>
     </div>
   )
 }
-
-// export const getServerSideProps = async () => {
-//   const loans = await fetchLoans()
-// const repayments = await fetchRepayments()
-
-//   return {
-//     props: {
-//       loans
-// repayments
-//     }
-//   }
-// }
